fix record aero-workout error

+++ crawler-web2/src/main/scala/workout/workoutDto.scala
@@ -22,7 +22,7 @@ case class UserAuth(
 
 @MongoDocument(databaseName="workout", collectionName="aerobicrecord")
 case class AerobicRecord(
-	usr: String, itm: String, tim: Int, dst: Int, clrs: Int, ts: Long) extends MongoModel
+	usr: String, itm: String, tim: Int, dst: Double, clrs: Int, ts: Long) extends MongoModel
 {
 	def this() = this(null, null, 0, 0, 0, 0)
 
@@ -34,7 +34,7 @@ case class AerobicRecord(
 	@MongoField var logTime = ts
 
 	override def toString = ("""AerobicRecord: {user: "%s",item: "%s",""" +
-		"time: %d, distance: %d, calories: %d, logTime: %d}").format(
+		"time: %d, distance: %f, calories: %d, logTime: %d}").format(
 		user, item, time, distance, calories, logTime)
 
 }
